Monitor is now a class and is instancied
