+++ modules/monitor/initializer.rb
@@ -1,4 +1,5 @@
 require_relative 'process'
 require_relative 'system'
+require_relative 'monitor'
 
 include Monitoring::System
